+++ tests/view_test/test_menu_view.py
@@ -14,13 +14,15 @@ def menubar(qtbot):
 def test_actions_attributes_exist(menubar):
     """Check that all QAction attributes are created on MainMenuBar."""
     assert isinstance(menubar.new_project_action, QAction)
+    assert isinstance(menubar.load_project_action, QAction)
     assert isinstance(menubar.load_file_action, QAction)
     assert isinstance(menubar.load_folder_action, QAction)
     assert isinstance(menubar.save_action, QAction)
     assert isinstance(menubar.exit_action, QAction)
 
     assert menubar.new_project_action.text() == "New Project"
+    assert menubar.load_project_action.text() == "Load Project"
     assert menubar.load_file_action.text() == "Load File(s)"
     assert menubar.load_folder_action.text() == "Load Folder"
-    assert menubar.save_action.text() == "Save"
+    assert menubar.save_action.text() == "Save Project"
     assert menubar.exit_action.text() == "Exit"
